requirements.txt
streamlit>=1.32.0
huggingface_hub>=0.21.0
python-dotenv>=1.0.0
pytest>=8.0.0
